After the creation of the project we have containerised it in the form of a docker image using a dockerfile. We then used a docker compose to create and run the containers.
Additionally we also focused on deployement in a production server. Which meant

1. Using whitenoise to server the static files to our server. we used collectstatic
to first fetch all of the static files. And then we used settings vars like staticfiles and
static_root to configure where the static files will be loaded from the server.

2. We used docker to containerise the application. The services communicate using container names inside
a docker network, so we use the service names in the settings.py file for  channel layer config.
Remember to switch to localhost for running it in local server as a uncontainerised application.

We use the gunicorn server for wsgi parts and a uvicorn worker for asgi part of the application as guncorn does not
support asgi application.